Struct aptos_vm::aptos_vm::AptosVM

pub struct AptosVM(_);

Implementations

impl AptosVM

pub fn new<S: StateView>(state: &S) -> Self

pub fn new_for_validation<S: StateView>(state: &S) -> Self

pub fn set_concurrency_level_once(concurrency_level: usize)

Sets execution concurrency level when invoked the first time.

pub fn get_concurrency_level() -> usize

Get the concurrency level if already set, otherwise return default 1 (sequential execution).

pub fn set_runtime_config(
paranoid_type_checks: bool,
paranoid_hot_potato_checks: bool
)

Sets runtime config when invoked the first time.

pub fn get_runtime_config() -> RuntimeConfig

Get the concurrency level if already set, otherwise return default true (paranoid execution mode).

pub fn set_num_proof_reading_threads_once(num_threads: usize)

Sets the # of async proof reading threads.

pub fn get_num_proof_reading_threads() -> usize

Returns the # of async proof reading threads if already set, otherwise return default value (32).

pub fn set_processed_transactions_detailed_counters()

Sets addigional details in counters when invoked the first time.

pub fn get_processed_transactions_detailed_counters() -> bool

Get whether we should capture additional details in counters

pub fn internals(&self) -> AptosVMInternals<'_>

pub fn load_module<S: MoveResolverExt>(
&self,
module_id: &ModuleId,
state: &S
) -> VMResult<Arc<CompiledModule>>

Load a module into its internal MoveVM’s code cache.

pub fn failed_transaction_cleanup<S: MoveResolverExt>(
&self,
error_code: VMStatus,
gas_meter: &mut AptosGasMeter,
txn_data: &TransactionMetadata,
storage: &S,
log_context: &AdapterLogSchema
) -> TransactionOutputExt

Generates a transaction output for a transaction that encountered errors during the execution process. This is public for now only for tests.

pub fn execute_block_and_keep_vm_status(
transactions: Vec<Transaction>,
state_view: &impl StateView
) -> Result<Vec<(VMStatus, TransactionOutput)>, VMStatus>

Alternate form of ‘execute_block’ that keeps the vm_status before it goes into the TransactionOutput

pub fn simulate_signed_transaction(
txn: &SignedTransaction,
state_view: &impl StateView
) -> (VMStatus, TransactionOutputExt)